outgoing call and voice message
There are activities showed under today's date in the calendar. One of them is a phone number. When the user clicks on it, a dial pad appears. When the call button is clicked a voice recording is played, when it ends, the user is returned to the calendar. The user can end the call beforehand and is also returned to the calendar.